requirements.txt
appdirs==1.4.4
asgiref==3.3.4
asn1crypto==1.4.0
attrs==21.2.0
audioread==2.1.9
awsebcli==3.20.0
bcrypt==3.2.0
bidict==0.21.2
blessed==1.18.1
boto3==1.17.60
botocore==1.20.60
cached-property==1.5.2
cachetools==4.2.2
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
cfgv==3.2.0
chardet==4.0.0
click==7.1.2
colorama==0.4.3
decorator==4.4.2
distlib==0.3.1
Django==3.2
django-allauth==0.44.0
django-bcrypt==0.9.2
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.1
django-filter==2.4.0
django-js-asset==1.2.2
django-json-widget==1.1.1
django-rest-auth==0.9.3
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.11.0
dnspython==1.16.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.20.0
enum-compat==0.0.3
eventlet==0.25.2
filelock==3.0.12
future==0.16.0
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-speech==2.3.0
google-cloud-texttospeech==2.3.0
googleapis-common-protos==1.53.0
greenlet==0.4.16
grpcio==1.37.1
hparams==0.3.0
identify==2.2.4
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
importlib-metadata==4.0.1
importlib-resources==5.1.2
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
libcst==0.3.18
librosa==0.8.0
llvmlite==0.36.0
MarkupSafe==1.1.1
monotonic==1.5
moviepy==1.0.3
mypy-extensions==0.4.3
nltk==3.5
nodeenv==1.6.0
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
openapi-codec==1.3.2
opencv-python==4.5.1.48
packaging==20.9
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.2.0
pooch==1.3.0
pre-commit==2.12.1
proglog==0.1.9
proto-plus==1.18.1
protobuf==3.15.8
py-bcrypt==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-engineio==4.1.0
python-rapidjson==1.0
python-socketio==5.2.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
scikit-learn==0.24.1
scikit-video==1.1.11
scipy==1.5.4
semantic-version==2.8.5
Shapely==1.7.1
simplejson==3.17.2
six==1.14.0
SoundFile==0.10.3.post1
sqlparse==0.4.1
termcolor==1.1.0
texttable==1.6.3
threadpoolctl==2.1.0
toml==0.10.2
torch==1.9.0
tqdm==4.60.0
tritonclient==2.9.0
typeguard==2.12.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
uritemplate==3.0.1
urllib3==1.25.11
virtualenv==20.4.4
wcwidth==0.1.9
websocket-client==0.59.0
wget==3.2
zipp==3.4.1